% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLDocument.R
\name{as.character.XMLDocument}
\alias{as.character.XMLDocument}
\title{Retrieve character representation of XMLDocument}
\usage{
\method{as.character}{XMLDocument}(x, include_declaration = TRUE, ...)
}
\arguments{
\item{x}{XMLDocument object}

\item{include_declaration}{Include the XML declaration at the top? Default: TRUE}

\item{...}{other arguments}
}
\description{
Retrieve character representation of XMLDocument
}
