% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLElement.R
\name{as.character.XMLElement}
\alias{as.character.XMLElement}
\title{Retrieve character representation of XMLElement}
\usage{
\method{as.character}{XMLElement}(x, include_declaration = FALSE, ...)
}
\arguments{
\item{x}{XMLElement object}

\item{include_declaration}{Include the XML declaration at the top? Default: FALSE}

\item{...}{other arguments}
}
\description{
Retrieve character representation of XMLElement
}
