% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_xml_doc}
\alias{parse_xml_doc}
\alias{parse_xml_elem}
\title{Parse xml text or file into an XMLDocument or XMLElement}
\usage{
parse_xml_doc(x, encoding = "", ..., as_html = FALSE,
  options = "NOBLANKS")

parse_xml_elem(x, encoding = "", ..., as_html = FALSE,
  options = "NOBLANKS")
}
\arguments{
\item{x, encoding, ..., as_html, options}{options passed to \code{xml2::read_xml()}}
}
\value{
XMLDocument or XMLElement
}
\description{
Parse xml text or file into an XMLDocument or XMLElement
}
