% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLDocument.R
\docType{data}
\name{XMLDocument}
\alias{XMLDocument}
\alias{xml_doc}
\title{XMLDocument Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XMLDocument
}
\description{
This is a slightly specialized subclass of \code{XMLElement}.
}
\details{
XML documents *must* have a root element
}
\keyword{datasets}
